db를 마이그레이션 하면서 남겨진 찌꺼기 파일 제거함.
